Rename to match logic to correct lambda stage
